Gas Delivery Service App Prototype
This gas service application was created by:
Jawad Townsend, Patrick Mills, and Alexander Whiteman.

This is an initial prototype of a gas delivery system application.
This service is designed so users are able to order gasoline to be
delivered while their car is parked. This system also has interfaces
for both administrators and fuel truck delivery drivers.